+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ```
+* Commit: 70992ff - Upload port-finder.js to GitHub Gist
+| Author: fjrodafo - Thu, 13 Feb 2025 02:49:55 +0000
+| 
+* Commit: f262843 - Update *.js comments
+| Author: fjrodafo - Thu, 13 Feb 2025 00:44:48 +0000
+| 
 * Commit: 123b1da - New VERSION.md
 | Author: fjrodafo - Thu, 13 Feb 2025 00:44:03 +0000
 |